python package reorganization to enable inclusion of edo parsers
