+++ configurations/ilc/ilc.config.file
@@ -1,15 +1,15 @@
  {
      "metadata": {
-        "equip_json": "/home/chandrika/volttron/tag-based-config-generators/sample_json/anacostia_equip_haystack.json",
-        "points_json": "/home/chandrika/volttron/tag-based-config-generators/sample_json/anacostia_points_haystack.json"
+        "equip_json": "/home/chandrika/volttron/tag-based-config-generators/sample_haystack_json/test_building_equip_haystack.json",
+        "points_json": "/home/chandrika/volttron/tag-based-config-generators/sample_haystack_json/test_building_points_haystack.json"
      },
-     "site_id": "r:@intellimation.dc_dgs.dcps.anacostia_hs",
+     "site_id": "r:@myorg.dc_dgs.dcps.test_building",
 
      # optional. if not provided will be derived from site_id.split('.')[-2]
      #"campus": "dcps",
 
      # optional. if not provided will be derived from site_id.split('.')[-1]
-     #"building": "anacostia_hs",
+     #"building": "test_building",
 
      ######### Configuration to identify whole building power meter. This required for ILC.  ##################
      ######### SHOULD MATCH the device name used by driver so that topic names match. ###########
@@ -27,7 +27,7 @@
      ##### 2. Use building power meter id. If there are multiple entities with the tag "siteMeter" then provide the
      ##### id of the power meter to use for ILC. id should exists in the tagging data store
 
-     "power_meter_id": "r:@intellimation.dc_dgs.dcps.anacostia_hs.electric_meter_-_building",
+     "power_meter_id": "r:@myorg.dc_dgs.dcps.test_building.electric_meter_-_building",
 
      ##### 3. When there is no power meter and agent is used to post data as powermeter - i.e. simulated power meter
      ##### specify both the "building_power_meter" and "building_power_point" attributes. In this case, config
@@ -42,7 +42,7 @@
             "WholeBuildingPower": "Mtr_kWh",
             "ZoneCoolingTemperatureSetPoint":  "EffZnTempSp",
             "MinimumAirFlow":"MinAirFlowSp",
-            #  "ZoneAirFlowSetpoint": "SaFlow", # Not used for anacostia. Derived as ZoneAirFlow*MaxAirFlow/100
+            #  "ZoneAirFlowSetpoint": "SaFlow", # Could be derived as ZoneAirFlow*MaxAirFlow/100
             "ZoneAirFlow": "SaFlow",
             "MaxAirFlow": "MaxAirFlowSP",
             "ZoneTemperature": "ZNTemp",
