+++ pyproject.toml
@@ -17,7 +17,7 @@ show_error_codes = true
 exclude = ['docs/']
 
 [tool.poetry]
-name = "volttron_config_gen"
+name = "volttron-config-generators"
 version = "0.1"
 description = "Parsers to generate VOLTTRON agent configurations based on haystack tags"
 authors = ["Chandrika Sivaramakrishnan <chandrika@pnnl.gov>"]
@@ -26,7 +26,7 @@ readme = "README.md"
 repository = "https://github.com/volttron/intellimation_config_generators"
 homepage = "https://github.com/volttron/intellimation_config_generators"
 keywords = []
-packages = [ { include = "volttron_config_gen", from = "src" } ]
+packages = [ { include = "volttron", from = "src" } ]
 classifiers = [
     "Programming Language :: Python :: 3 :: Only",
     "Intended Audience :: Science/Research",
