+++ sample_haystack_json/test_building_points_haystack.json
