+++ src/volttron/config_gen/base/config_economizer.py
@@ -3,12 +3,12 @@ import copy
 import os.path
 import sys
 from abc import abstractmethod
-from volttron_config_gen.haystack.parser.utils import strip_comments
+from volttron.config_gen.utils import strip_comments
 
 
 class AirsideEconomizerConfigGenerator:
     """
-    Base class that parses haystack tags to generate
+    Base class that parses haystack3 tags to generate
     Airside Economizer agent configuration based on a configuration template
     """
     def __init__(self, config):
