+++ src/volttron/config_gen/haystack3/intellimation/config_ilc.py
@@ -1,22 +1,19 @@
-import copy
-import json
-import re
 import sys
 import psycopg2
 
-from volttron_config_gen.haystack.parser.ilc.config_base import ILCConfigGenerator
+from volttron.config_gen.base.config_ilc import ILCConfigGenerator
 
 
 class IntellimationILCConfigGenerator(ILCConfigGenerator):
     """
-    class that parses haystack tags from a postgres db to generate
+    class that parses haystack3 tags from a postgres db to generate
     ILC configs
     """
 
     def __init__(self, config):
         super().__init__(config)
 
-        # get details on haystack metadata
+        # get details on haystack3 metadata
         metadata = self.config_dict.get("metadata")
         connect_params = metadata.get("connection_params")
         if "timescale_dialect" in metadata:
